Conflicts:
	metricbeat/docker-entrypoint.sh
	metricbeat/docs/fields.asciidoc
	metricbeat/etc/beat.yml
	metricbeat/etc/fields.yml
	metricbeat/etc/fields_base.yml
	metricbeat/metricbeat.template-es2x.json
	metricbeat/metricbeat.template.json
	metricbeat/metricbeat.yml
	metricbeat/module/nginx/*

+++ topbeat/system/cpu.go
@@ -10,8 +10,7 @@ import (
 )
 
 type CPU struct {
-	CpuPerCore bool
-
+	CpuPerCore       bool
 	LastCpuTimes     *CpuTimes
 	LastCpuTimesList []CpuTimes
 }
@@ -31,7 +30,6 @@ func GetCpuTimes() (*CpuTimes, error) {
 	}
 
 	return &CpuTimes{Cpu: cpu}, nil
-
 }
 
 func GetCpuTimesList() ([]CpuTimes, error) {
@@ -78,7 +76,7 @@ func GetCpuPercentageList(last, current []CpuTimes) []CpuTimes {
 		calculate := func(field2 uint64, field1 uint64, all_delta uint64) float64 {
 
 			perc := 0.0
-			delta := field2 - field1
+			delta := int64(field2 - field1)
 			perc = float64(delta) / float64(all_delta)
 			return Round(perc, .5, 4)
 		}
